+++ ignite/pkg/gocmd/gocmd_test.go
@@ -12,7 +12,7 @@ import (
 func TestIsInstallError(t *testing.T) {
 	assert.False(t, gocmd.IsInstallError(errors.New("oups")))
 
-	err := errors.New(`error while running command go install github.com/regen-network/cosmos-proto/protoc-gen-gocosmos github.com/golang/protobuf/protoc-gen-go github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2: no required module provides package github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2; to add it:
+	err := errors.New(`error while running command go install github.com/cosmos/gogoproto/protoc-gen-gocosmos github.com/golang/protobuf/protoc-gen-go github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2: no required module provides package github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2; to add it:
 		go get github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2`)
 	assert.True(t, gocmd.IsInstallError(err))
 }