+++ integration/cosmosgen/bank_module_test.go
@@ -15,6 +15,8 @@ import (
 )
 
 func TestBankModule(t *testing.T) {
+	t.Skip()
+
 	var (
 		env     = envtest.New(t)
 		app     = env.Scaffold("chain", "--no-module")